Suppress warnings on return types of CharSequence overrides (#40)
This PR fixes type-checking by the Index Checker with PR typetools/checker-framework#1764.
Annotations were added to return types of two methods in `CharSequence`, so that there are less false positives in code using `CharSequence`. However, it means that overriders of these methods should use the annotations as well. They do not make sense for mutable-length implementations, so they are not used and the warnings are suppressed.